CI testing via GitLab (#248)
* gitlab support
* disable dialog when no tty
* split builds into stages
* add vmware wrapper
* fix for dovecot port rename
* avg: add -y when no tty
* switch pkg from quarterly -> latest
* prevent variable interpolation
* IP whitelist the staging & monitor servers"
* reduce busy fs timeout
* vpopmail: use unique nrpe function name

+++ provision-monitor.sh
@@ -34,7 +34,11 @@ install_munin()
 
 configure_munin()
 {
-	rm -r "$STAGE_MNT/usr/local/etc/munin"
+	if [ -d "$STAGE_MNT/data/etc/munin" ]; then
+		rm -r "$STAGE_MNT/usr/local/etc/munin"
+	else
+		mv "$STAGE_MNT/usr/local/etc/munin" "$STAGE_MNT/data/etc/"
+	fi
 	stage_exec ln -s /data/etc/munin /usr/local/etc/munin
 
 	stage_sysrc munin_node_enable=YES
